library/pyproject.toml
[tool.poetry]
name = "tradingcore"
version = "0.1.0"
description = ""
authors = ["KBM15 <keyblademaster_01@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
numpy = "1.26.4"
yfinance = "^0.2.40"
pandas-ta = "^0.3.14b0"
datetime = "^5.5"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.9"
fastapi = "^0.112.1"
uvicorn = "^0.30.6"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
